requirements.txt
fastapi==0.65.2
uvicorn==0.12.2

pydantic~=1.8.2
starlette~=0.14.2
SQLAlchemy~=1.3.22
pytest==6.2.2
bcrypt~=3.2.0
PyJWT~=2.0.0